+++ src/steps/IRAS/reviewResearch/userAdministration/manageReviewBodies/CreateReviewBodySteps.ts
@@ -1,10 +1,7 @@
 import { createBdd } from 'playwright-bdd';
 import { expect, test } from '../../../../../hooks/CustomFixtures';
-import { generateTimeStampedValue, generateUniqueValue } from '../../../../../utils/UtilFunctions';
+import { generateTimeStampedValue } from '../../../../../utils/UtilFunctions';
 import { Locator } from 'playwright/test';
-import path from 'path';
-const pathToTestDataJson =
-  './src/resources/test_data/iras/reviewResearch/userAdministration/manageReviewBodies/create_review_body_page_data.json';
 
 const { When, Then } = createBdd(test);
 
@@ -63,27 +60,6 @@ Then(
       await expect(createReviewBodyPage.email_address_error).toHaveText(dataset.Mandatory_Field);
     } else {
       await expect(createReviewBodyPage[fieldName.toLowerCase()]).toHaveText(dataset[errorMsg]);
-    }
-  }
-);
-
-When(
-  'I fill the new review body page for audit using {string}',
-  async ({ createReviewBodyPage, commonItemsPage }, datasetName: string) => {
-    const dataset = createReviewBodyPage.createReviewBodyPageData.Create_Review_Body[datasetName];
-    for (const key in dataset) {
-      if (Object.prototype.hasOwnProperty.call(dataset, key)) {
-        if (key === 'organisation_name_text') {
-          const prefix = createReviewBodyPage.createReviewBodyPageData.Create_Review_Body.organisation_name_prefix;
-          const uniqueOrgName = await generateUniqueValue(dataset[key], prefix);
-          const filePath = path.resolve(pathToTestDataJson);
-          await createReviewBodyPage.updateUniqueOrgNameTestDataJson(filePath, uniqueOrgName);
-          const locator = createReviewBodyPage[key];
-          await locator.fill(uniqueOrgName);
-        } else {
-          await commonItemsPage.fillUIComponent(dataset, key, createReviewBodyPage);
-        }
-      }
     }
   }
 );
